+++ buildSrc/src/Versions.kt
@@ -1,10 +1,11 @@
-package com.serebit.strife.gradle
+package com.serebit.strife.buildsrc
 
 /** Versions of dependencies for type-safe consistency. */
 object Versions {
-    const val COROUTINES = "1.3.0-RC"
-    const val KTOR = "1.2.3"
-    const val SERIALIZATION = "0.11.1"
-    const val LOGKAT = "0.4.6"
-    const val KLOCK = "1.5.0"
+    const val COROUTINES = "1.3.1"
+    const val KTOR = "1.2.4"
+    const val SERIALIZATION = "0.12.0"
+    const val LOGKAT = "0.4.7"
+    const val KLOCK = "1.6.2"
+    const val JUPITER = "5.5.2"
 }