+++ client/src/commonMain/kotlin/BotDsl.kt
@@ -1,8 +1,7 @@
 package com.serebit.strife
 
-import com.serebit.strife.events.Event
-import com.serebit.strife.events.MessageCreatedEvent
-import com.serebit.strife.events.ReadyEvent
+import com.serebit.strife.events.*
+import com.serebit.strife.internal.EventResult
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
@@ -44,23 +43,52 @@ inline fun CoroutineScope.launchBot(token: String, crossinline init: BotBuilder.
 }
 
 /**
- * Creates an event listener for events with type T. The code inside the [task] block will be executed every time
- * the bot receives an event with type T.
+ * Creates a terminable event listener of type [T], which will run the given [task] [successfulRunLimit] number of
+ * [successful][EventResult.SUCCESS] times before being terminated.
+ */
+@BotBuilderDsl
+@UseExperimental(ExperimentalStdlibApi::class)
+inline fun <reified T : Event> BotBuilder.onEventLimited(
+    successfulRunLimit: Int = 1,
+    noinline task: suspend T.() -> EventResult
+) = onTerminableEvent(typeOf<T>(), successfulRunLimit, task)
+
+/**
+ * Creates an event listener for events with type [T]. The code inside the [task] block will be executed every time the
+ * bot receives an event with type [T].
  */
 @BotBuilderDsl
 @UseExperimental(ExperimentalStdlibApi::class)
 inline fun <reified T : Event> BotBuilder.onEvent(noinline task: suspend T.() -> Unit) = onEvent(typeOf<T>(), task)
 
-/** Convenience method to create an event listener that will execute on reception of a ReadyEvent. */
+/** Convenience method to create an event listener that will execute when the bot starts a session. */
 @BotBuilderDsl
-@UseExperimental(ExperimentalStdlibApi::class)
-fun BotBuilder.onReady(task: suspend ReadyEvent.() -> Unit) {
-    onEvent(typeOf<ReadyEvent>(), task)
-}
+fun BotBuilder.onReady(task: suspend ReadyEvent.() -> Unit) = onEvent(task)
+
+/** Convenience method to create an event listener that will execute when the bot resumes a session. */
+@BotBuilderDsl
+fun BotBuilder.onResume(task: suspend ResumeEvent.() -> Unit) = onEvent(task)
 
 /** Convenience method to create an event listener that will execute when a message is created. */
 @BotBuilderDsl
-@UseExperimental(ExperimentalStdlibApi::class)
-fun BotBuilder.onMessage(task: suspend MessageCreatedEvent.() -> Unit) {
-    onEvent(typeOf<MessageCreatedEvent>(), task)
-}
+fun BotBuilder.onMessageCreate(task: suspend MessageCreateEvent.() -> Unit) = onEvent(task)
+
+/** Convenience method to create an event listener that will execute when a message's content is edited. */
+@BotBuilderDsl
+fun BotBuilder.onMessageEdit(task: suspend MessageEditEvent.() -> Unit) = onEvent(task)
+
+/** Convenience method to create an event listener that will execute when a message is deleted. */
+@BotBuilderDsl
+fun BotBuilder.onMessageDelete(task: suspend MessageDeleteEvent.() -> Unit) = onEvent(task)
+
+/** Convenience method to create an event listener that will execute when a channel is created. */
+@BotBuilderDsl
+fun BotBuilder.onChannelCreate(task: suspend ChannelCreateEvent.() -> Unit) = onEvent(task)
+
+/** Convenience method to create an event listener that will execute when a channel is updated. */
+@BotBuilderDsl
+fun BotBuilder.onChannelUpdate(task: suspend ChannelUpdateEvent.() -> Unit) = onEvent(task)
+
+/** Convenience method to create an event listener that will execute when a channel is deleted. */
+@BotBuilderDsl
+fun BotBuilder.onChannelDelete(task: suspend ChannelDeleteEvent.() -> Unit) = onEvent(task)