+++ client/src/commonMain/kotlin/internal/entitydata/GuildData.kt
@@ -1,6 +1,7 @@
 package com.serebit.strife.internal.entitydata
 
 import com.serebit.strife.BotClient
+import com.serebit.strife.RemoveCacheData
 import com.serebit.strife.data.Presence
 import com.serebit.strife.data.toPermissions
 import com.serebit.strife.data.toPresence
@@ -33,12 +34,11 @@ internal class GuildData(
 
     val channelList: Collection<GuildChannelData<*, *>> get() = channels.values
 
-    private val roles = packet.roles.asSequence()
+    val roles = packet.roles.asSequence()
         .map { context.cache.pullRoleData(it) }
+        .onEach { it.guildId = id }
         .associateBy { it.id }
         .toMutableMap()
-
-    val roleList: Collection<GuildRoleData> get() = roles.values
 
     private var emojis = packet.emojis.asSequence()
         .map { context.cache.pullEmojiData(this, it) }
@@ -141,7 +141,7 @@ internal class GuildData(
 
     fun update(data: GuildRoleDelete.Data) {
         roles.remove(data.role_id)
-        context.cache.removeRoleData(data.role_id)
+        context.cache.remove(RemoveCacheData.GuildRole(data.role_id))
     }
 
     fun update(data: GuildEmojisUpdate.Data) {