+++ samples/ping/src/Ping.kt
@@ -1,9 +1,9 @@
 package samples
 
-import com.serebit.strife.bot
+import com.serebit.strife.*
 import com.serebit.strife.entities.reply
-import com.serebit.strife.onMessage
-import com.serebit.strife.onReady
+import com.serebit.strife.events.MessageCreateEvent
+import com.serebit.strife.internal.EventResult
 
 /**
  * An example of how to use Strife to connect
@@ -15,19 +15,14 @@ suspend fun main(args: Array<String>) {
 
     // Start the bot building scope
     bot(token) {
-         logToConsole = true // Uncomment this to see log messages
+        logToConsole = true // Remove or set this to false to hide log messages
 
         // Print to console when the bot is connected & ready
         onReady { println("Connected to Discord!") }
 
         // On "!ping" messages, send PONG!
-        onMessage {
-            if (message.content == "!ping") {
-                message.reply("Pong! :ping_pong: ")?.also {
-                    // After the message is replied to, edit it to show the delay
-                    it.edit("Pong! :ping_pong: ${(it.createdAt - message.createdAt).millisecondsLong}ms")
-                }
-            }
+        onMessageCreate {
+            if (message.content == "!ping") message.reply("Pong! :ping_pong:")
         }
     }
 }