+++ CHANGELOG.md
@@ -3,27 +3,54 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
-### Added
+## Unreleased
+
+#### Added 
+- More shorthand BotBuilder extensions such as `onChannelCreate`, `onMessageEdit`, etc
+- Releases are now mirrored to Maven Central
+- `GuildRole` now has a `guildId` property, alongside `getGuild` and `delete` functions
+- [#17](https://gitlab.com/serebit/strife/issues/17) Add global markdown extensions on String
+- [#20](https://gitlab.com/serebit/strife/issues/20) Add terminable event listeners, which are removed from the client when their task is successful
+
+#### Changed 
+- Rename some events for consistency
+- Rename some BotBuilder extensions for consistency
+
+## 0.2.1 (2019-09-03)
+
+#### Changed 
+- Updated stack to Kotlin 1.3.50
+
+#### Fixed 
+- Doubled the speed of encoding bytes to base64
+
+## 0.2.0 (2019-08-12)
+
+#### Added 
 - Add a new `Presence` class to support all presence properties
 - Activity now supports showing thumbnails for Spotify songs
 - Implement `GuildRoleCreateEvent`, `GuildRoleUpdateEvent`, `GuildRoleDeleteEvent`, `GuildIntegrationsUpdateEvent` and `GuildMembersChunkEvent`
 - Add `presence` property to `GuildMember`
 - Add `getChannel`, `getTextChannel`, `getVoiceChannel`, and `getRole` functions to `Guild`
 - Add `presences` property to `Guild`
-### Changed
+
+#### Changed 
 - Optimizations for Color math
 - Improve existing `Activity` class and integrate it into `Presence`
 - `BotClient.updatePresence()` no longer uses `Activity`, instead it uses `Pair<Activity.Type, String>`
 - `PresenceUpdateEvent` now provides the full `Presence` object instead of only `Activity` and `OnlineStatus`
 - `PresenceUpdateEvent` now implements `GuildEvent`
-### Fixed
-- Fix roles and channels not being updated in Guild
 
-## [0.1.2] - 2019-08-02
-### Added
+#### Fixed 
+- Fix roles and channels not being updated in Guild
+- Fix memory leaking when sending HTTP requests
+
+## 0.1.2 (2019-08-02)
+
+#### Added 
 - Add Guild.getMember() to allow getting a member by their ID. This will attempt to get the member from Discord API if it's not found in the cache
-### Changed
+
+#### Changed 
 - Updated to Kotlin 1.3.41
 - All guild dispatches will now wait for the guild to become available before being dispatched
 - Improved the internal representation of entity classes and how events update them
@@ -32,20 +59,25 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Replace the Java platform base-64 encoder with our own implementation
 - Replace Guild.owner with Guild.getOwner(), in case the owner member is no longer in the cache
 - Members are no longer be cached permanently, and will be removed from the cache if going unused
-### Fixed
+
+### Fixed 
 - [#14](https://gitlab.com/serebit/strife/issues/14) Add Watching activity type
 
-## [0.1.1] - 2019-06-30
-### Added
+## 0.1.1 (2019-06-30)
+
+#### Added 
 - Greatly improved logging for events that could not be processed correctly. The log message now shows which event type failed to process and why the failure happened
-### Changed
+
+#### Changed 
 - Updated to Kotlin 1.3.40
 - Changed `UnicodeEmoji.Companion::invoke` to a method with name `fromUnicode`
 - `fromUnicode` now checks if the passed unicode is a valid unicode emoji
-### Removed
+
+#### Removed 
 - `!stop` command has been removed from the `ping` sample
 - Removed transitive dependency on kotlin-reflect, along with any reflection usages in strife-jvm
-### Fixed
+
+#### Fixed 
 - The process now properly exits when `BotClient::disconnect` is called
 - `is` operator now works correctly with UnicodeEmoji
 - Deserializing guilds no longer fails due to `joined_at` missing